Fix type of $timeout parameter of JsonClient::receive
Php::Type::Numeric is an integer, not a float.
See:
- https://github.com/yaroslavche/phptdlib/blob/0.0.10/tdlib.cpp#L52
- https://github.com/CopernicaMarketingSoftware/PHP-CPP/blob/v2.2.0/include/type.h#L25

+++ stubs/TDLib/JsonClient.php
@@ -5,10 +5,10 @@ class JsonClient extends BaseJsonClient {
     private $defaultTimeout = 0.5;
 
     /**
-     * @param float|null $timeout
+     * @param int|null $timeout
      * @return string
      */
-    public function receive(float $timeout = null):string
+    public function receive(int $timeout = null):string
     {
     }
